chore: move util file from plugin to component
