+++ packages/vchart/src/component/tooltip/utils/pattern.ts
@@ -1,6 +1,6 @@
-import type { IDimensionInfo } from '../../../../event/events/dimension/interface';
-import type { ISeries } from '../../../../series/interface';
-import type { ITooltipPattern, TooltipActiveType } from '../../../../typings';
+import type { IDimensionInfo } from '../../../event/events/dimension/interface';
+import type { ISeries } from '../../../series/interface';
+import type { ITooltipPattern, TooltipActiveType } from '../../../typings';
 
 /**
  * 根据图表上下文生成默认的tooltip pattern
