+++ packages/vchart/src/plugin/components/tooltip-handler/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './common';
-export * from './pattern';
+export * from '../../../../component/tooltip/utils/pattern';
 export * from './position';
 export * from './attribute';
